change delete category action for POST method,

+++ src/AppBundle/Controller/CategoryController.php
@@ -10,6 +10,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
 use AppBundle\Form\Type\CategoryType;
+use AppBundle\Form\Type\YesNoType;
 
 
 
@@ -71,20 +72,22 @@ class CategoryController extends Controller
      */
     public function DelCategoryAction(Request $request, Category $category)
     {
-        return $this->render('category/categoryDelete.html.twig', array('category' => $category) );
-    }
+        $form = $this->createForm(new YesNoType());
 
-    /**
-     * @Route("/category/delete/accepted/{id}", name="accDelCat")
-     */
-    public function DeleteCategoryAction(Request $request, Category $category) {   
-        
-        $form = $this->createForm(new CategoryType(), $category);
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($category);
-        $entityManager->flush(); 
-   
-        return $this->redirect($this->generateUrl('listCat'));
+        if($form->handleRequest($request)->isValid()) {
+            if($form->get('tak')->isClicked()) {
+                $form = $this->createForm(new CategoryType(), $category);
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($category);
+                $entityManager->flush(); 
+           
+                return $this->redirect($this->generateUrl('listCat'));                
+            } else {
+                return $this->redirect($this->generateUrl('listCat'));                
+            }
+        }
+
+        return $this->render('category/categoryDelete.html.twig', array('category' => $category, 'form' => $form->createView()));
     }
 }
